Got location detection om image load working.
For the library we are using images should be loaded from same domain.

Placed a symlink to images dir under webapp and for tomcat to resolve it
added context.xml where conf param to allowsymlinks is set. _Events
script configures tomcat configuration on startup.

To undo this effect change rsources.serverURL and observations.serverURL
to prev params.